backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.19
pydantic[email]==2.7.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1